requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
SQLAlchemy==2.0.46
alembic==1.14.1
psycopg2-binary==2.9.10
python-jose==3.5.0
pydantic==2.12.5
firebase-admin==6.7.0
requests==2.32.5
pytest==8.3.5
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.0.1
